+++ lib/ruby_sage/engine.rb
@@ -19,7 +19,19 @@ module RubySage
     initializer "ruby_sage.assets" do |app|
       assets_config = app.config.assets if app.config.respond_to?(:assets)
 
-      assets_config.precompile += %w[ruby_sage_manifest.js] if assets_config.respond_to?(:precompile)
+      if assets_config.respond_to?(:precompile)
+        assets_config.precompile += %w[
+          ruby_sage/widget.css
+          ruby_sage/widget.js
+          ruby_sage_manifest.js
+        ]
+      end
+    end
+
+    initializer "ruby_sage.helpers" do
+      ActiveSupport.on_load(:action_controller_base) do
+        helper RubySage::WidgetHelper
+      end
     end
   end
 end