+++ spec/dummy/config/routes.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
+  root "posts#index"
+
   resources :posts, only: %i[index show]
 
   mount RubySage::Engine => "/ruby_sage"